+++ app/controllers/roles_controller.rb
@@ -2,6 +2,8 @@ class RolesController < ApplicationController
   before_action :load_role, only: [:show, :update, :edit, :destroy]
 
   def index
+    authorize :role, :index?
+
     @roles = Role.all
     respond_to do |f|
       f.html
@@ -9,6 +11,8 @@ class RolesController < ApplicationController
     end
   end
   def show
+    authorize @role
+
     respond_to do |f|
       f.html
       f.json { render json: @role }
@@ -16,6 +20,8 @@ class RolesController < ApplicationController
   end
 
   def update
+    authorize @role
+
     if @role.update(role_params)
       redirect_to @role
     else
@@ -23,10 +29,13 @@ class RolesController < ApplicationController
     end
   end
   def edit
+    authorize @role
   end
 
   def create
     @role = Role.new(role_params)
+    authorize @role
+
     if @role.save
       redirect_to @role
     else
@@ -35,9 +44,12 @@ class RolesController < ApplicationController
   end
   def new
     @role = Role.new
+    authorize @role
   end
 
   def destroy
+    authorize @role
+
     if @role.destroy
       redirect_to root_path
     else
