+++ app/controllers/users_controller.rb
@@ -2,6 +2,8 @@ class UsersController < ApplicationController
   before_action :load_user, only: [:show, :update, :edit, :destroy, :new_role, :destroy_role]
 
   def index
+    authorize :user, :index?
+
     @users = User.all
     respond_to do |f|
       f.html
@@ -9,6 +11,8 @@ class UsersController < ApplicationController
     end
   end
   def show
+    authorize @user
+
     @roles = @user.roles
 
     respond_to do |f|
@@ -18,6 +22,8 @@ class UsersController < ApplicationController
   end
 
   def update
+    authorize @user
+
     if @user.update(user_params)
       redirect_to @user
     else
@@ -29,6 +35,8 @@ class UsersController < ApplicationController
 
   def create
     @user = User.new(user_params)
+    authorize @user
+
     if @user.save
       redirect_to @user
     else
@@ -37,13 +45,18 @@ class UsersController < ApplicationController
   end
   def new
     @user = User.new
+    authorize @user
   end
   def new_role
+    authorize @user
+
     UsersRole.create user: @user, role_id: params[:user][:role_id]
     redirect_to @user
   end
 
   def destroy
+    authorize @user
+
     if @user.destroy
       redirect_to root_path
     else
@@ -51,6 +64,8 @@ class UsersController < ApplicationController
     end
   end
   def destroy_role
+    authorize @user
+
     UsersRole.destroy_all(user_id: @user, role_id: params[:role_id])
     redirect_to @user
   end
